Add confirmation to apply-configs

+++ apply-configs.sh
@@ -2,7 +2,12 @@
 
 REPO_DOTFILES_DIR="$HOME/davidrv-dotfiles/dotfiles"
 
-for f in $(cat $REPO_DOTFILES_DIR/.dotfiles); do
-    mkdir -pv $HOME/$(dirname "$f") &&
-    cp -rTv $REPO_DOTFILES_DIR/$f $HOME/$f;
-done
+read -p "This will overwrite system config files. Are you sure? (Y/y to confirm): " -n 1 -r
+echo
+if [[ $REPLY =~ ^[Yy]$ ]]
+then
+  for f in $(cat $REPO_DOTFILES_DIR/.dotfiles); do
+      mkdir -pv $HOME/$(dirname "$f") &&
+      cp -rTv $REPO_DOTFILES_DIR/$f $HOME/$f;
+  done
+fi
